Update Harfbuzz to 7.2.0
Qt 5.15 has been lagging behind on Harfbuzz because of
compile issues. This is an attempt at doing a full upgrade,
duplicating most of the modifications we have done along the
way.

Two major differences from Qt 6: We still use hb-coretext for
macOS shaping, since changing this is not really within scope
for the current commit policy of the LTS. In addition, we
enable deprecated API because we cannot change the minimum
supported system-harfbuzz version, and thus we cannot rely on
the replacement APIs.

There are still older platforms which cannot compile the new
Harfbuzz. Rather than including an outdated third party library
for these platforms, we disable Harfbuzz-NG and use Qt 4's
built-in shaper instead (equivalent to QT_HARFBUZZ=old).
Users can still use -system-harfbuzz on these platforms if they
have local versions of Harfbuzz that build.

This adds a QSKIP to tst_qfontmetrics::zeroWidthMetrics. Looking
at the code which introduced that, it primarily fixes it for
freetype. Apparently it ends up passing with Harfbuzz-NG on
Windows because we resolve it to a different glyph and this
ends up with a large right bearing and negative width. But
since it depends on the metrics of the corresponding glyph, it
may not be very reliable in this case either. At least for the
case where we use the old harfbuzz, the weakness is exposed and
the test fails, so for now we just skip the test on this
configuration.

[ChangeLog][Important behavior change] The included version of
the text shaping library Harfbuzz has been disabled on a selection
of older platforms where the updated versions do not compile. There
may be regressions in text appearance with certain fonts when
rendering languages that depend on OpenType features due to this.
A possible mitigation is to build an older Harfbuzz version
manually and use -system-harfbuzz to use this with Qt.
This affects the following platforms: QNX, Integrity, WinRT,
Intel ICC compiler, MacOS 10.15 and lower, MSVC 2015 and lower.

Task-number: QTBUG-113352
Task-number: QTBUG-111350
Change-Id: I0fad6841c41018586b21c6e2dcd4166c26ba5716
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 6b92ef96e965a64b9a3e92f8ea3b2f3a57897ef7)