Darwin: Cache system locale's zero digit and invalidate on locale changes
As part of ed2b110b6add650954dc102a0317c14ff826c677 the implicit caching
of some of the system locale queries was removed, resulting in a
three orders of magnitude performance regression on converting
dates to strings via QLocale::system() on macOS and iOS.

We now cache the most critical of these, the zero digit, in the Darwin
backend, and clear the cache in response to the system locale changing,
restoring the performance two orders of a magnitude closer to that of Qt 5.

We're still one order of magnitude off, which seems to stem from the
auto release pool in QSystemLocale::query().

Fixes: QTBUG-104785
Change-Id: I26f96147d430f08a721c55a1048d586a4af3a76c
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit d3ede38eddf94b24d195ed58e13182117d7e2b74)
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>