Fix case-sensitivity of exponent separator check in Cyrillic fall-back
This is a back-port to 6.5 of both the special-case handling of
Cyrillic's exponent separator [0] and the fix to that, that prompted
this cherry-pick. Both changes were made after significant
restructuring of relevant code to cope with multi-character exponent
separators, which is not backported to 6.5.

[0] commit cb54da236662563bfdc2f58a01320c3bd360cbff

Added some test-cases for the lower-case exponential separator.

Fixes: QTBUG-113443
Change-Id: I18e22d7b3451fbb61e87d5b93661eadff3c7356e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>
(cherry picked from commit 4768fcf836f1eb0f59ab62a29eb821b433a239c1)
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>